pyproject.toml
# Copyright 2020 The Hyp2Rem Authors

# Use of this source code is governed by an MIT-style
# license that can be found in the LICENSE file or at
# https://opensource.org/licenses/MIT.


[tool.poetry]
name = "hyp2rem"
version = "0.1.0"
description = "Hypothes.is to RemNote integration"
authors = ["The Hyp2Rem Authors"]
license = "MIT"
readme = "README.md"
exclude = ["**/~BROMIUM"]

[tool.poetry.scripts]
hyp2rem = "hyp2rem.__main__:app"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.0"
minilog = "^2.0"
typer = {extras = ["all"], version = "^0.3.2"}
python-dotenv = "^0.15.0"
pylru = "^1.2.0"
pydantic = "^1.7.3"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
black = "^20.8b1"
isort = "^5.6.4"
mypy = "^0.790"
pre-commit = "^2.9.0"
pylint = "^2.6.0"

[tool.pylint.messages_control]
disable = [
    "too-many-arguments",
    "too-many-instance-attributes",
    "fixme",
    # only until pylint fixes typing support for Python3.9
    "unsubscriptable-object",
]

[tool.pytest.ini_options]
minversion = "6.1"
addopts = "-ra -q"
testpaths = [
    "tests"
]
log_cli = true
log_cli_level = "INFO"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

